requirements.txt
beautifulsoup4==4.12.2
boto3==1.26.165
botocore==1.29.165
cotk==0.1.0
embeddings==0.0.8
faiss_cpu==1.7.4
Flask==3.0.0
flask_admin==1.6.1
flask_limiter==3.5.0
flask_socketio==5.3.6
gensim==4.3.2
intel_extension_for_transformers==1.2.2
jieba==0.42.1
jsonlines==3.1.0
langchain==0.0.351
nltk==3.8.1
numpy==1.23.5
pandas==2.0.2
peewee==3.17.0
peft==0.4.0.dev0
ptvsd==4.3.2
pyarrow==12.0.0
pyserini==0.22.1
python-dotenv==1.0.0
python_bcrypt==0.3.2
pytorch_pretrained_bert==0.6.2
quadprog==0.1.11
redis_lock==0.2.0
Requests==2.31.0
scipy==1.11.4
setuptools==69.0.2
setuptools==59.6.0
six==1.16.0
tensorboardX==2.6.2.2
tensorflow==2.15.0.post1
torch==2.0.1
torch_geometric==2.4.0
tqdm==4.65.0
transformers==4.34.1
